==> results/01_LogsToCSV.ipynb <==
#%reset
import pandas as pd
from pandas import DataFrame, Series
#false positive warnings all the time:
pd.options.mode.chained_assignment = None

%matplotlib inline
import matplotlib as mpl
import matplotlib.pyplot as plt

import numpy as np
import statistics

import os.path

from datetime import datetime


import re

#source: http://stackoverflow.com/questions/5917082/regular-expression-to-match-numbers-with-or-without-commas-and-decimals-in-text
regex_number_with_commas = '(\d+|\d{1,3}(,\d{3})*)(\.\d+)?'
regex_decimal_number = '(\d+\.)?\d+'


def removeCommasInteger(i):
    return int(i.replace(",",""))

#Thanks to: https://regex101.com/ for testing the regexes live!

# 1. Simulation identifiers and paths

Simulation is identified as a **tuple**: (Store, Number of nodes, Amount of RAM, Dataset, Config, Additional info field)

example_tuple = ("Fuseki", 1, 64, "Ontoforce", "Default", "")

## Helper functions

* simulation tuple -> long_identifier
* simulation tuple -> short_identifier
* simulation tuple -> path to log file


def generateSimulationID(tup):
    identifier = tup[0] \
            + "_N" + str(tup[1]) \
            + "_" + str(tup[2]) \
            + "_" + tup[3] \
            + "_" + tup[4] \
    
    extra = ""
    if len(tup[5]) !=0:
        extra = "_" + str(tup[5])
    
    return identifier+extra


simulation_map = {
    "Watdiv10M": "W10", "Watdiv100M": "W100", "Watdiv1000M": "W1000", "Ontoforce": "Ont"
}

def generateSimulationIDCompact(tup, pref_length=3):
    #pref_length = 3
    identifier = tup[0][:pref_length] \
            + "_N" + str(tup[1]) \
            + "_" + str(tup[2]) \
            + "_" + simulation_map[tup[3]] \
            + "_" + str(tup[4])[:pref_length] \
    
    extra = ""
    if len(tup[5]) !=0:
        extra = "_" + str(tup[5])
    
    return identifier+extra

def generateFullPathErrorLog(path, tup):
    return path + tup[0] + '/' +  generateSimulationID(tup) + "_output.log"

### TEST

home_path = "./"

print(generateSimulationID(example_tuple))
print(generateSimulationIDCompact(example_tuple))
sim_identifier = generateSimulationIDCompact(example_tuple)
full_path = generateFullPathErrorLog(home_path,example_tuple)
print(full_path)
print('Log file found? ' + str(os.path.isfile(full_path))) 

# 2. Convert log file into csv document

Fields:

- query_order: the index of the query
- query_name
- threadID
- number of results
- runtime
- simulation_id

## 2.1 Extract Query Thread IDs

def getIDsWarmupThreads(path):

    queryThreadList = []
    querying_started = False

    with open(path) as f:
        for line in f:

            if 'Running Warmups...' in line:
                querying_started = True
            
            if 'Running Benchmarks...' in line:
                break
                
            if not querying_started:
                continue

            
            if 'Running Operation ' in line:
                m = re.search('Thread (\d+)',line)
                queryThreadList.append(m.group(1))

    return list(set(queryThreadList))

def getIDsQueryThreads(path):
    queryThreadList = []
    querying_started = False
    
    with open(path) as f:
        for line in f:
            
            if 'Running Benchmarks...' in line:
                querying_started = True
            
            #some of the log files are appended versions of multiple benchmark runs
            if querying_started and 'Running Warmups...' in line:
                break
                
            if not querying_started:
                continue
                
            if 'Running Operation ' in line:
                m = re.search('Thread (\d+)',line)
                queryThreadList.append(m.group(1))

    return list(set(queryThreadList))

### TEST

ids_warmupthreads = getIDsWarmupThreads(full_path)
print(ids_warmupthreads)
ids_querythreads = getIDsQueryThreads(full_path)
print(ids_querythreads)

## 2.2 Extract logs per queryID

def getLogLinesOfWThread(qid, path):
    lines = []
    querying_started = False
    
    c = 0
    c_start = 0
    
    with open(path) as f:
        for line in f:
            c+=1
            
            if 'Running Warmups...' in line:
                c_start = c
                querying_started = True
            
            
            if 'Running Benchmarks...' in line:
                querying_started = False
                
                break
              
            if not querying_started:
                continue
            
            if ('Thread ' + str(qid) + "]") in line:
                
                if 'Running Operation' in line:
                    if '...[Thread' in line: #loglines accidentally concatenated
                        line1 = line[:line.find('...[Thread') +3]
                        line2 = line[line.find('...[Thread') +3:]
                        lines.append(line1.strip())
                        lines.append(line2.strip())
                        
                        #print(line1)
                        #print(line2)
                    else:
                        lines.append(line.strip())
                elif 'result(s) in' in line:
                    lines.append(line.strip())
                elif 'got error' in line:
                    lines.append(line.strip())
                    
    #print("Loglines: " + str(c_start) + " -> " + str(c))
    return lines

def getLogLinesOfQThread(qid, path):
    lines = []
    querying_started = False

    c = 0
    c_start = 0
    
    with open(path) as f:
        for line in f:
            c+=1
            
            if 'Running Benchmarks...' in line:
                querying_started = True
                c_start = c
            
            #some of the log files are appended versions of multiple benchmark runs
            if querying_started and 'Running Warmups...' in line:
                break
                
            if not querying_started:
                continue
            
            if ('Thread ' + str(qid) + "]") in line:
                
                if 'Running Operation' in line:
                    if '...[Thread' in line: #loglines accidentally concatenated
                        line1 = line[:line.find('...[Thread') +3]
                        line2 = line[line.find('...[Thread') +3:]
                        lines.append(line1.strip())
                        lines.append(line2.strip())
                        
                        #print(line1)
                        #print(line2)
                    else:
                        lines.append(line.strip())
                elif 'result(s) in' in line:
                    lines.append(line.strip())
                elif 'got error' in line:
                    lines.append(line.strip())

    #print("Loglines: " + str(c_start) + " -> " + str(c))

    return lines

### TEST

for tid in ids_warmupthreads:
    print(str(tid) + ":\t" + str(len(getLogLinesOfWThread(tid, full_path))))
for tid in ids_querythreads:
    print(str(tid) + ":\t" + str(len(getLogLinesOfQThread(tid, full_path))))    
    
lines = getLogLinesOfWThread(1, full_path)


## 2.3 Convert logs per thread into query information tuples

Log lines consist of a line stating a query is executed (contains query name) and a second line stating whether the query succeeded, the number of results and the runtime. Note that these lines are with the same thread and not consecutively in the log file!

def generateTupleFromResultsLine(line, sim_id, query_name, thread_id, thread_type, order_id):
    
    regex_results = 'got (' + regex_number_with_commas + ') result'
    regex_seconds = 'in ('+ regex_decimal_number + ')s$'
    
    m = re.search(regex_results, line)
    number_of_results = removeCommasInteger(m.group(1))
    m = re.search(regex_seconds, line)
    number_of_seconds = float(m.group(1))
    
    return (sim_id, query_name, thread_id, thread_type, order_id, number_of_results, number_of_seconds, 'SUCCESS')

example_line = lines[15]
print(example_line)
print(generateTupleFromResultsLine(example_line, generateSimulationIDCompact(example_tuple), \
                                  'query_name', 1, 'warmup', 0))

def generateQueryEventList(full_path, tid, thread_type, queryPath, timeout, sim_id):
    resultList = []
    
    lines = None
    if thread_type == 'warmup':
        lines = getLogLinesOfWThread(tid, full_path)
    else :
        lines = getLogLinesOfQThread(tid, full_path)
    
    readResults = False
    query = None
    c = 0
    
    for l in lines:
        
   
        if readResults:
            

            if 'result(s) in ' in l:
                
                tup = generateTupleFromResultsLine(l, sim_id, query, tid, thread_type, c)
                resultList.append(tup)
                

            elif 'got error' in l:

                if 'Operation Callable' in l:

                    tup = (sim_id, query, tid, thread_type, c, -1, timeout, 'TIMEOUT')
                    resultList.append(tup)
                    
                else:
                    #print(l)
                    tup = (sim_id, query, tid, thread_type, c, -1, -1, 'ERROR')
                    resultList.append(tup)                                    
            else:

                #if next line is also a 'Running Operation' then we also flag as error
                tup = (sim_id, query, tid, thread_type, c, -1, -1, 'ERROR')
                resultList.append(tup)                                    


            readResults = False

        if 'Running Operation' in l:

            p = l.find(queryPath)
            p_end = l.find('...')
            
            if p>=0: 
                readResults = True
                query = str(l[p+len(queryPath):-3])
                #print(query)
                c+=1
    
    return resultList

### TEST

eventlist = generateQueryEventList(full_path, 1, 'warmup', "templated/", 300, sim_identifier)
c=0
for e in eventlist:
   
    print(e)
    c+=1
    if c==10:
        break

## 2.4 Create dataframe with tuples of 1 thread

#WARNING: don't append row by row, it's very slow, internally creates a new dataframe since dataframes are immutable
# speed went from +3 minutes -> subsecond!!

def queryEventsToDataframe(queryEventList):
    
    infomaps = []
    c=0
    for tup in queryEventList:
        
        infomap = {}
        infomap['sim_id'] = tup[0]
        infomap['query_name'] = tup[1]
        infomap['thread_id'] = int(tup[2])
        infomap['thread_type'] = tup[3]
        infomap['order_id'] = int(tup[4])
        infomap['number_of_results'] = int(tup[5])
        infomap['runtime'] = tup[6]
        infomap['flag'] = tup[7]
        row = pd.Series(infomap)
        #df.loc[c] = row
        c+=1
        infomaps.append(row)
        
        #df.convert_objects(convert_numeric=True)
        
    df = pd.concat(infomaps, axis=1).T
    df = df[['sim_id', 'query_name', 'thread_id', 'thread_type', 'order_id', 'number_of_results', \
                               'runtime', 'flag']]
    return df

### TEST

df = queryEventsToDataframe(eventlist)
df.head(n=5)

## 2.5 Merge dataframes of all threads

def generateQueryEventDataframe(bm_tuple, home_path, query_path, timeout):
    
    path_logfile = generateFullPathErrorLog(home_path, bm_tuple)
    sim_id = generateSimulationIDCompact(bm_tuple)
    print("PROCESSING SIM_ID: " + sim_id)
    print(path_logfile)
    
    found = os.path.isfile(path_logfile)
    print('Log file found? ' + str(found))
    
    if found is False:
        return False
    
    warmup_t = getIDsWarmupThreads(path_logfile)
    query_t = getIDsQueryThreads(path_logfile)

    print(warmup_t)
    print(query_t)
    
    df_tot = None
    
    for tid in warmup_t:
        #loglines = getLogLinesOfThread(tid, path_logfile)
        #print("WThread "+str(tid)+":\t" + str(len(loglines)) + " loglines found")
        eventlist = generateQueryEventList(path_logfile, tid, 'warmup', query_path, timeout, sim_id)
        print("WThread " + str(tid) + ": " + str(len(eventlist)) + " query events found")
        
        #TODO remove
        print(eventlist[0])
        
        if df_tot is not None:
            df = queryEventsToDataframe(eventlist)
            df_tot = pd.concat([df_tot, df])
        else:
            df_tot = queryEventsToDataframe(eventlist)
            
    for tid in query_t:
        #loglines = getLogLinesOfThread(tid, path_logfile)  
        #print("QThread "+str(tid)+":\t" + str(len(loglines)) + " loglines found")
        eventlist = generateQueryEventList(path_logfile, tid, 'stress', query_path, timeout, sim_id)
        print("QThreads " + str(tid)+ ": " + str(len(eventlist)) + " query events found")
        
        if df_tot is not None:
            df = queryEventsToDataframe(eventlist)
            df_tot = pd.concat([df_tot, df])
        else:
            df_tot = queryEventsToDataframe(eventlist)
            

    
    return df_tot
    

### TEST

df_tot = generateQueryEventDataframe(example_tuple, './', 'templated/', 300)
df_tot.shape

df_tot.head(n=5)

## 2.6 Save to CSV 

filename = sim_identifier + "_queryevents.csv"
df_tot.to_csv('./csv/'+filename, sep="\t", header=True, index=False)

# 3 Convert all simulation data to csv

## A. Watdiv (10M, 100M, 1000M)


#(1, 32, "Watdiv10M", "Default", ""), \

#(1, 32, "Watdiv100M", "Default", ""), \
#(1, 64, "Watdiv100M", "Default", ""), \
#(3, 64, "Watdiv100M", "Default", ""), \
#
#(1, 32, "Watdiv1000M", "Default", ""), \
#(1, 64, "Watdiv1000M", "Default", ""), \
#(1, 64, "Watdiv1000M", "Optimized", "")]    
#(3, 32, "Watdiv1000M", "Optimized", "")]    

### 1. Blazegraph

bm_tuples = [ ("Blazegraph", 1, 32, "Watdiv10M", "Default", ""), \
("Blazegraph", 1, 32, "Watdiv100M", "Default", ""), \
("Blazegraph", 1, 32, "Watdiv1000M", "Default", ""), \
("Blazegraph", 1, 64, "Watdiv1000M", "Default", ""), \
("Blazegraph", 1, 64, "Watdiv1000M", "Optimized", "")]

for bm_tuple in bm_tuples:
    
    print(str(datetime.now()))
    
    df = generateQueryEventDataframe(bm_tuple, './', 'templated/', 300)

    if df is not False:
        filename = generateSimulationID(bm_tuple)+ "_queryevents.csv"
        df.to_csv('./csv/'+filename, sep="\t", header=True, index=False)
print(str(datetime.now()))

### 2. GraphDB

bm_tuples = [ ("GraphDB", 1, 32, "Watdiv10M", "Default", ""), \
("GraphDB", 1, 32, "Watdiv100M", "Default", ""), \
("GraphDB", 1, 32, "Watdiv1000M", "Default", ""), \
("GraphDB", 1, 64, "Watdiv1000M", "Default", ""), \
("GraphDB", 1, 64, "Watdiv1000M", "Optimized", "")]

for bm_tuple in bm_tuples:
    
    print(str(datetime.now()))
    
    df = generateQueryEventDataframe(bm_tuple, './', 'templated/', 300)
    
    if df is not False:
        filename = generateSimulationID(bm_tuple)+ "_queryevents.csv"
        df.to_csv('./csv/'+filename, sep="\t", header=True, index=False)

print(str(datetime.now()))

### 3. ES

bm_tuples = [ ("ES", 1, 32, "Watdiv10M", "Default", ""), \
("ES", 1, 32, "Watdiv100M", "Default", ""), \
("ES", 1, 32, "Watdiv1000M", "Default", ""), \
("ES", 3, 32, "Watdiv1000M", "Default", ""), \
("ES", 1, 64, "Watdiv1000M", "Default", "")]

for bm_tuple in bm_tuples:
    
    print(str(datetime.now()))
    
    df = generateQueryEventDataframe(bm_tuple, './', 'templated/', 300)
    
    if df is not False:
        filename = generateSimulationID(bm_tuple)+ "_queryevents.csv"
        df.to_csv('./csv/'+filename, sep="\t", header=True, index=False)

print(str(datetime.now()))

### 4. Virtuoso

bm_tuples = [ ("Virtuoso", 1, 32, "Watdiv10M", "Default", ""), \
("Virtuoso", 1, 32, "Watdiv100M", "Default", ""), \
("Virtuoso", 1, 32, "Watdiv1000M", "Default", ""), \
("Virtuoso", 1, 32, "Watdiv1000M", "Default", "RERUN"), \
("Virtuoso", 3, 32, "Watdiv1000M", "Default", ""), \
("Virtuoso", 1, 64, "Watdiv1000M", "Default", ""), \
("Virtuoso", 1, 64, "Watdiv1000M", "Optimized", "")]

for bm_tuple in bm_tuples:
    
    print(str(datetime.now()))
    
    df = generateQueryEventDataframe(bm_tuple, './', 'templated/', 300)
    
    if df is not False:
        filename = generateSimulationID(bm_tuple)+ "_queryevents.csv"
        df.to_csv('./csv/'+filename, sep="\t", header=True, index=False)

print(str(datetime.now()))

### 5. Fuseki

bm_tuples = [ ("Fuseki", 1, 64, "Watdiv100M", "Default", ""), \
("Fuseki", 1, 64, "Watdiv1000M", "Default", "")]

for bm_tuple in bm_tuples:
    
    print(str(datetime.now()))
    
    df = generateQueryEventDataframe(bm_tuple, './', 'templated/', 300)
    
    if df is not False:
        filename = generateSimulationID(bm_tuple)+ "_queryevents.csv"
        df.to_csv('./csv/'+filename, sep="\t", header=True, index=False)

print(str(datetime.now()))

### 6. LDF

bm_tuples = [ ("LDF", 1, 64, "Watdiv100M", "Default", ""), \
("LDF", 3, 64, "Watdiv100M", "Default", ""), \
("LDF", 1, 64, "Watdiv1000M", "Default", ""), \
("LDF", 3, 64, "Watdiv1000M", "Default", "")]

for bm_tuple in bm_tuples:
    
    print(str(datetime.now()))
    
    df = generateQueryEventDataframe(bm_tuple, './', 'templated/', 300)
    
    if df is not False:
        filename = generateSimulationID(bm_tuple)+ "_queryevents.csv"
        df.to_csv('./csv/'+filename, sep="\t", header=True, index=False)

print(str(datetime.now()))

### 7. FluidOps

bm_tuples = [ ("FluidOps", 3, 64, "Watdiv100M", "Default", ""), \
("FluidOps", 1, 64, "Watdiv1000M", "Default", ""), \
("FluidOps", 3, 64, "Watdiv1000M", "Default", "")]
 
for bm_tuple in bm_tuples:
    
    print(str(datetime.now()))
    
    df = generateQueryEventDataframe(bm_tuple, './', 'templated/', 300)
    
    if df is not False:
        filename = generateSimulationID(bm_tuple)+ "_queryevents.csv"
        df.to_csv('./csv/'+filename, sep="\t", header=True, index=False)

print(str(datetime.now()))

## B. Ontoforce Dataset

bm_tuples = [ ("Blazegraph", 1, 64, "Ontoforce", "Optimized", ""), \
("GraphDB", 1, 64, "Ontoforce", "Optimized", ""), \
("ES", 1, 64, "Ontoforce", "Default", ""), \
("Virtuoso", 1, 64, "Ontoforce", "Optimized", ""), \
("Virtuoso", 1, 64, "Ontoforce", "Optimized", "VWall"), \
("Virtuoso", 1, 32, "Ontoforce", "Optimized", ""), \
("Virtuoso", 1, 32, "Ontoforce", "Optimized", "VWall"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "0"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "1"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "2"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "AWS1"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "AWS2"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "AWS3"), \
("Fuseki", 1, 64, "Ontoforce", "Default", ""), \
("FluidOps", 1, 64, "Ontoforce", "Default", "1"), \
("FluidOps", 1, 64, "Ontoforce", "Default", "2"), \
("FluidOps", 1, 64, "Ontoforce", "Default", "3"), \
("FluidOps", 3, 64, "Ontoforce", "Default", "1"), \
("FluidOps", 3, 64, "Ontoforce", "Default", "2"), \
("FluidOps", 3, 64, "Ontoforce", "Default", "3")]
             

for bm_tuple in bm_tuples:
    
    print(str(datetime.now()))
    
    df = generateQueryEventDataframe(bm_tuple, './', 'templated/', 1200)
    
    if df is not False:
        filename = generateSimulationID(bm_tuple)+ "_queryevents.csv"
        df.to_csv('./csv/'+filename, sep="\t", header=True, index=False)

print(str(datetime.now()))
