==> results/16_QueriesThatCrashSimulations.ipynb <==
import pandas as pd
from pandas import DataFrame, Series
#false positive warnings all the time:
pd.options.mode.chained_assignment = None

%matplotlib inline
import matplotlib as mpl
import matplotlib.pyplot as plt

import numpy as np
import statistics

import os.path

#http://matplotlib.org/1.4.0/users/customizing.html

plt.style.use('ggplot')

mpl.rcParams['font.size'] = 18

### LINES
mpl.rcParams['lines.linewidth'] = '2.0'        # line width in points
mpl.rcParams['lines.marker']    = 'None'
mpl.rcParams['lines.markersize']=  '6'          # markersize, in points

###LEGEND
mpl.rcParams['legend.fontsize']='medium'

### GRIDS
mpl.rcParams['grid.color']='black'     # grid color
mpl.rcParams['grid.linestyle'] =':'    # dotted
mpl.rcParams['grid.linewidth']='0.5'   # in points
mpl.rcParams['grid.alpha']='1.0'       # transparency, between 0.0 and 1.0

### TICKS
mpl.rcParams['xtick.labelsize']="small"  # fontsize of the tick labels
mpl.rcParams['xtick.direction']="out"      # direction: in, out, or inout
mpl.rcParams['ytick.labelsize']="small"   # fontsize of the tick labels
mpl.rcParams['ytick.direction']="out"      # direction: in, out, or inout

### AXES
mpl.rcParams['axes.linewidth']  =  '1.0'    # edge linewidth
mpl.rcParams['axes.grid']       =  'True'   # display grid or not
mpl.rcParams['axes.titlesize']  =  'large'  # fontsize of the axes title
mpl.rcParams['axes.labelsize']  = 'large'  # fontsize of the x any y labels
mpl.rcParams['axes.facecolor']= 'white'     # axes background color

mpl.rcParams['figure.facecolor']= '1.0'     # figure facecolor; 0.75 is scalar gray

def generateSimulationID(tup):
    identifier = tup[0] \
            + "_N" + str(tup[1]) \
            + "_" + str(tup[2]) \
            + "_" + tup[3] \
            + "_" + tup[4] \
    
    extra = ""
    if len(tup[5]) !=0:
        extra = "_" + str(tup[5])
    
    return identifier+extra


simulation_map = {
    "Watdiv10M": "W10", "Watdiv100M": "W100", "Watdiv1000M": "W1000", "Ontoforce": "Ont"
}

def generateSimulationIDCompact(tup, pref_length=3):
    #pref_length = 3
    identifier = tup[0][:pref_length] \
            + "_N" + str(tup[1]) \
            + "_" + str(tup[2]) \
            + "_" + simulation_map[tup[3]] \
            + "_" + str(tup[4])[:pref_length] \
    
    extra = ""
    if len(tup[5]) !=0:
        extra = "_" + str(tup[5])
    
    return identifier+extra

def generate_sorted_events_filename(tup):
    return 'query_events_sorted/'+generateSimulationID(tup)+"_events_sorted.csv"

#crashes on watdiv
#fuseki watdiv
#fluidops watdiv

tuple_list = [
    ('Fuseki', 1, 64,'Watdiv100M', 'Default', ''), \
    ('FluidOps', 3, 64,'Watdiv100M', 'Default', ''), \
    ('FluidOps', 3, 64,'Watdiv1000M', 'Default', ''), \
    ('Fuseki', 1, 64,'Watdiv1000M', 'Default', '')
    
]

tup = tuple_list[0]
df = pd.read_csv(generate_sorted_events_filename(tup), sep='\t', index_col=0)
df.head(n=3)

mask1 = df['flag'] == 'SUCCESS'
mask2 = df['correct'] == 'CORRECT'
df_filtered = df[mask1 & mask2]
df_filtered.groupby('thread_id')[['order_id']].max()

df_filtered.groupby('thread_id')[['order_id']].max().sum()

rows = []
rows.append(df[(df['thread_id'] == 101) & (df['order_id'] == 44)])
rows.append(df[(df['thread_id'] == 103) & (df['order_id'] == 51)])
rows.append(df[(df['thread_id'] == 104) & (df['order_id'] == 59)])
rows.append(df[(df['thread_id'] == 105) & (df['order_id'] == 53)])
rows.append(df[(df['thread_id'] == 107) & (df['order_id'] == 53)])

pd.concat(rows)

df_filtered['template'] = df_filtered['query_name'].apply(lambda q: q[:2])
df_filtered2 = df_filtered[df_filtered['thread_type'] == 'stress']
pd.DataFrame(df_filtered2['template'].value_counts()).sort_index()

* Fuseki handles all queries in the warmup phase, but crashes on the first C2 query in the stress test!

tup = tuple_list[1]
df = pd.read_csv(generate_sorted_events_filename(tup), sep='\t', index_col=0)
df.head(n=3)

mask1 = df['flag'] == 'SUCCESS'
mask2 = df['correct'] == 'CORRECT'
df_filtered = df[mask1 & mask2]
df_filtered.groupby('thread_id')[['order_id']].max()

rows = []
rows.append(df[(df['thread_id'] == 1) & (df['order_id'] == 41)])


pd.concat(rows)

df_filtered['template'] = df_filtered['query_name'].apply(lambda q: q[:2])
pd.DataFrame(df_filtered['template'].value_counts()).sort_index()

* FluidOps crashes on the first C3 query!!!

tup = tuple_list[2]
df = pd.read_csv(generate_sorted_events_filename(tup), sep='\t', index_col=0)
df.head(n=3)

mask1 = df['flag'] == 'SUCCESS'
mask2 = df['correct'] == 'CORRECT'
df_filtered = df[mask1 & mask2]
df_filtered.groupby('thread_id')[['order_id']].max()

rows = []
rows.append(df[(df['thread_id'] == 1) & (df['order_id'] == 5)])


pd.concat(rows)

df.head(n=5)

* FluidOps crashes on the first C3 query!!!

tup = tuple_list[3]
df = pd.read_csv(generate_sorted_events_filename(tup), sep='\t', index_col=0)
df.head(n=3)

mask1 = df['flag'] == 'SUCCESS'
mask2 = df['correct'] == 'CORRECT'
df_filtered = df[mask1 & mask2]
df_filtered.groupby('thread_id')[['order_id']].max()

rows = []
rows.append(df[(df['thread_id'] == 1) & (df['order_id'] == 52)])


pd.concat(rows)

df_filtered['template'] = df_filtered['query_name'].apply(lambda q: q[:2])
pd.DataFrame(df_filtered['template'].value_counts()).sort_index()

* Fuseki crashes on the first C2 query

# Summary

* Fuseki Watdiv100M
    - 655 queries before crash, 400 = warmup successful, approx 50 queries per thread in stress test
    - 16u15 until crash
    - first C2 query in stress test generates HTTP error, all other templates succeeded
* Fuseki Watdiv1000M
    - 51 queries before crash, doesn't survive warmup phase
    - 2u35 until crash
    - first C2 query in warmup test generates HTTP error, all other templates succeeded
* FluidOps N3 Watdiv100M
    - 40 queries before crash, doesn't survive warmup phase!
    - 2013 seconds until crash
    - first C3 query in warmup results in TIMEOUT, other templates succeeded! From then on no more successes
* FluidOps N3 Watdiv1000M
    - 4 queries before crash, doesn't survive warmup phase!
    - 572 seconds until crash
    - first C3 query in warmup results in TIMEOUT, other templates succeeded! Fron then on no more successes
