==> results/18_RuntimeAnalysisCSV_DiscardIncorrectQueries.ipynb <==
import pandas as pd
from pandas import DataFrame, Series
#false positive warnings all the time:
pd.options.mode.chained_assignment = None

%matplotlib inline
import matplotlib as mpl
import matplotlib.pyplot as plt

import numpy as np
import statistics

import os.path

import time
def print_elapsed(start):
    t = time.time() - start
    print(str(t) + " secs")

# Runtime Analysis

## Methods:

* Assumption: queries are correct! This assumption will be invalidated for a number of simulations later on!

### A1. Define Tquery?

* Constraint: only look for query runtimes during the time the server was up (=benchmark survival). There are cases where the server crashed and the benchmarker reported timeouts.

* Distinguish between time in the warmup phase (single threaded) and the stress test (multithreaded)

* Per query report the number successful runs, the number of error runs and the number of timeouts

* For the median runtime take timeouts and successful runtimes into account

* **NEW** Remove incorrect queries before 

* **NEW** Remove C3 templates (pd.merge will handle this as C3s won't be in the index for Virtuoso

* If query is never seen during benchmark survival we cannot say anything => it's not in our data and plots!


### A2. Tquery dataframe:

**PREFILTER ON BENCHMARK SURVIVAL**

** IF LOG FILE NOT AVAILABLE:** use median reported in results csv file with thread_type=stress, successes=all, errors=0, timeouts=0

**COLUMNS:**

- thread_type: warmup/stress
- successes: number of times flag == SUCCESS
- errors: number of times flag == ERROR
- timeouts: number of times flag == TIMEOUT
- incorrect: number of times correct = INCORRECT




## 1. Query Runtime dataframe from events file: Tryout

def generateFilenameQueryEventsCorrect(bmtuple):
    fname = './csv_correct/' +  bmtuple[0] + "_N" + str(bmtuple[1]) + "_" + str(bmtuple[2]) + "_" + bmtuple[3] + "_" + bmtuple[4]
    if len(str(bmtuple[5])) != 0:
        fname = fname + "_" + str(bmtuple[5])
    return fname + "_queryevents_correct.csv"

def generateFilenameRuntimesCorrect(bmtuple):
    fname = './runtime_csv_correct/' +  bmtuple[0] + "_N" + str(bmtuple[1]) + "_" + str(bmtuple[2]) + "_" + bmtuple[3] + "_" + bmtuple[4]
    if len(str(bmtuple[5])) != 0:
        fname = fname + "_" + str(bmtuple[5])
    return fname + "_queryruntimes_correct.csv"

def generate_sorted_events_filename(tup):
    return 'query_events_sorted/'+generateSimulationID(tup)+"_events_sorted.csv"

### TEST

events_file = generateFilenameQueryEventsCorrect(('Virtuoso',1,32,'Watdiv1000M', 'Default',''))
print(events_file)
print(os.path.isfile(events_file))

df = pd.read_csv(events_file, sep='\t')
df.head()

### Max Order ID per thread

df_filter_success = df[df['flag'] == 'SUCCESS']
df_survival = df_filter_success.groupby(['thread_id', 'thread_type'], as_index=False)[['sim_id','order_id']].max()

for line in df_survival.values:
    print(line)


### For a thread get the corresponding restriced df

df_test = df[(df['thread_id'] == 16) & (df['thread_type'] == 'stress' )]  
df_test2 = df_test[df_test['order_id'] <= 82] 
df_test2.head()

def preFilterForBMSurvival(filename):
    df = pd.read_csv(filename, sep='\t')
    
    df_filter_success = df[df['flag'] == 'SUCCESS']
    df_survival = df_filter_success.groupby(['thread_id', 'thread_type'], as_index=False)[['sim_id','order_id']].max()

    
    restricted_dfs = []
    for tup in df_survival.values:
        thread_id = tup[0]
        thread_type = tup[1]
        df_thread = df[(df['thread_id'] == thread_id) & (df['thread_type'] == thread_type )]  
        
        max_order_id = tup[3]
        df_thread_restricted = df_thread[df_thread['order_id'] <= max_order_id]
        restricted_dfs.append(df_thread_restricted)
    
    return pd.concat(restricted_dfs, axis=0)
        
     
    
        
    

### TEST 

df_test = preFilterForBMSurvival(generateFilenameQueryEventsCorrect(('Virtuoso',1,32,'Watdiv1000M', 'Default','')))

print(df_test.shape)
print(12000)

### timeouts, errors, successes column values

TIMEOUT = df_test['flag'] == 'TIMEOUT'
ERROR   = df_test['flag'] == 'ERROR'
SUCCESS = df_test['flag'] == 'SUCCESS'
CORRECT = df_test['correct'] == 'CORRECT'

df_timeouts = df_test [TIMEOUT]
df_errors = df_test [ERROR]
df_successes = df_test [SUCCESS & CORRECT]
df_incorrect = df_test [SUCCESS & ~CORRECT]

df_nonerrors =  pd.concat([df_timeouts, df_successes])

print(df_timeouts.shape)
print(df_errors.shape)
print(df_successes.shape)
print(df_incorrect.shape)

print(df_nonerrors.shape)

ntimeouts = df_timeouts.groupby(['query_name', 'thread_type'], as_index=False).agg({'flag': np.size})
ntimeouts = ntimeouts.rename(columns={'flag': 'timeouts'})
ntimeouts.head()

nerrors = df_errors.groupby(['query_name', 'thread_type'], as_index=False).agg({'flag': np.size})
nerrors = nerrors.rename(columns={'flag': 'errors'})
nerrors.head()

nsuccess = df_successes.groupby(['query_name', 'thread_type'], as_index=False).agg({'flag': np.size})
nsuccess = nsuccess.rename(columns={'flag': 'success'})
nsuccess.head()

nincorrect = df_incorrect.groupby(['query_name', 'thread_type'], as_index=False).agg({'flag': np.size})
nincorrect = nincorrect.rename(columns={'flag': 'incorrect'})
nincorrect.head()

### runtime nonerror queries

nonerror_timings = df_nonerrors.groupby(['query_name', 'thread_type'], as_index=False).agg({'runtime': np.median})
nonerror_timings = nonerror_timings.rename(columns={'runtime': 'median_runtime'})
nonerror_timings.tail()

### merge together

df_merge1 = pd.merge(ntimeouts, nerrors, how='outer', on=['query_name','thread_type'])
df_merge1 = df_merge1.fillna(0)
df_merge2 = pd.merge(df_merge1, nsuccess, how='outer', on=['query_name','thread_type'])
df_merge2 = df_merge2.fillna(0)

#LEAVE OUT: incorrect queries are unusable for runtime comparisons!

#df_merge3 = pd.merge(df_merge2, nincorrect, how='outer', on=['query_name','thread_type'])
#df_merge3 = df_merge3.fillna(0)
df_merge3 = df_merge2

df_merge4 = pd.merge(df_merge3, nonerror_timings, how='outer', on=['query_name','thread_type'])
timeout = 300
df_merge4 = df_merge4.fillna(timeout)
df_merge4.head()

#print(df_merge4.columns)
df_merge5 = df_merge4[['query_name', 'thread_type', 'errors', 'timeouts', 'success', 'median_runtime']]

df_merge5.to_csv('test_tquery.csv', sep="\t", index=False)

## 2. Query Runtime for all available files

def createQueryRuntimeDF(filename, querytimeout):
    df_survival = preFilterForBMSurvival(filename)
        
    TIMEOUT = df_survival['flag'] == 'TIMEOUT'
    ERROR   = df_survival['flag'] == 'ERROR'
    SUCCESS = df_survival['flag'] == 'SUCCESS'
    CORRECT = df_survival['correct'] == 'CORRECT'

    df_timeouts  = df_survival [TIMEOUT]
    df_errors    = df_survival [ERROR]
    df_successes = df_survival [SUCCESS & CORRECT]
    df_incorrect = df_survival [SUCCESS & ~CORRECT]

    df_nonerrors =  pd.concat([df_timeouts, df_successes])
    

    
    
    ntimeouts        = df_timeouts.groupby(['query_name', 'thread_type'], as_index=False).agg({'flag': np.size})
    nerrors          = df_errors.groupby(['query_name', 'thread_type'], as_index=False).agg({'flag': np.size})
    nsuccess         = df_successes.groupby(['query_name', 'thread_type'], as_index=False).agg({'flag': np.size})
    nincorrect       = df_incorrect.groupby(['query_name', 'thread_type'], as_index=False).agg({'flag': np.size})
    nonerror_timings = df_nonerrors.groupby(['query_name', 'thread_type'], as_index=False).agg({'runtime': np.median})
    
    ntimeouts        = ntimeouts.rename(columns={'flag': 'timeouts'})
    nerrors          = nerrors.rename(columns={'flag': 'errors'})
    nsuccess         = nsuccess.rename(columns={'flag': 'success'})
    nincorrect       = nincorrect.rename(columns={'flag': 'incorrect'})
    nonerror_timings = nonerror_timings.rename(columns={'runtime': 'median_runtime'})
 
    df_merge1 = pd.merge(ntimeouts, nerrors, how='outer', on=['query_name','thread_type'])
    df_merge1 = df_merge1.fillna(0)
    df_merge2 = pd.merge(df_merge1, nsuccess, how='outer', on=['query_name','thread_type'])
    df_merge2 = df_merge2.fillna(0)

    #LEAVE OUT: incorrect queries are unusable for runtime comparisons!

    #df_merge3 = pd.merge(df_merge2, nincorrect, how='outer', on=['query_name','thread_type'])
    #df_merge3 = df_merge3.fillna(0)
    df_merge3 = df_merge2

    df_merge4 = pd.merge(df_merge3, nonerror_timings, how='outer', on=['query_name','thread_type'])
    df_merge4 = df_merge4.fillna(querytimeout)
    
    return df_merge4[['query_name', 'thread_type', 'success', 'errors', 'timeouts', 'median_runtime']]



### 1. Blazegraph Watdiv


def runtimes_for_stores(tuples, timeout):
    start = time.time()
    
    for bm_tuple in bm_tuples:
    
        filename = generateFilenameQueryEventsCorrect(bm_tuple)

        print(filename + ": " + str(os.path.isfile(filename)))

        if os.path.isfile(filename):

            df = createQueryRuntimeDF(filename, timeout)
            new_filename = generateFilenameRuntimesCorrect(bm_tuple)
            df.to_csv(new_filename, sep="\t", header=True, index=False)

    print_elapsed(start)
    return

bm_tuples = [ ("Blazegraph", 1, 32, "Watdiv10M", "Default", ""), \
("Blazegraph", 1, 32, "Watdiv100M", "Default", ""), \
("Blazegraph", 1, 32, "Watdiv1000M", "Default", ""), \
("Blazegraph", 1, 64, "Watdiv1000M", "Default", ""), \
("Blazegraph", 1, 64, "Watdiv1000M", "Optimized", "")]

querytimeout = 300
runtimes_for_stores(bm_tuples, querytimeout)


### 2. GraphDB Watdiv

bm_tuples = [ ("GraphDB", 1, 32, "Watdiv10M", "Default", ""), \
("GraphDB", 1, 32, "Watdiv100M", "Default", ""), \
("GraphDB", 1, 32, "Watdiv1000M", "Default", ""), \
("GraphDB", 1, 64, "Watdiv1000M", "Default", ""), \
("GraphDB", 1, 64, "Watdiv1000M", "Optimized", "")]

runtimes_for_stores(bm_tuples, querytimeout)


### 3. ES Watdiv

bm_tuples = [ ("ES", 1, 32, "Watdiv10M", "Default", ""), \
("ES", 1, 32, "Watdiv100M", "Default", ""), \
("ES", 1, 32, "Watdiv1000M", "Default", ""), \
("ES", 3, 32, "Watdiv1000M", "Default", ""), \
("ES", 1, 64, "Watdiv1000M", "Default", "")]

runtimes_for_stores(bm_tuples, querytimeout)


### 4. Virtuoso Watdiv

bm_tuples = [ ("Virtuoso", 1, 32, "Watdiv10M", "Default", ""), \
("Virtuoso", 1, 32, "Watdiv100M", "Default", ""), \
("Virtuoso", 1, 32, "Watdiv1000M", "Default", ""), \
("Virtuoso", 1, 32, "Watdiv1000M", "Default", "RERUN"), \
("Virtuoso", 3, 32, "Watdiv1000M", "Default", ""), \
("Virtuoso", 1, 64, "Watdiv1000M", "Default", ""), \
("Virtuoso", 1, 64, "Watdiv1000M", "Optimized", "")]

runtimes_for_stores(bm_tuples, querytimeout)



### 5. Fuseki Watdiv

bm_tuples = [ ("Fuseki", 1, 64, "Watdiv100M", "Default", ""), \
("Fuseki", 1, 64, "Watdiv1000M", "Default", "")]

runtimes_for_stores(bm_tuples, querytimeout)


### 6. LDF Watdiv

bm_tuples = [ ("LDF", 1, 64, "Watdiv100M", "Default", ""), \
("LDF", 3, 64, "Watdiv100M", "Default", ""), \
("LDF", 1, 64, "Watdiv1000M", "Default", ""), \
("LDF", 3, 64, "Watdiv1000M", "Default", "")]

runtimes_for_stores(bm_tuples, querytimeout)


### 7. FluidOps Watdiv

bm_tuples = [ ("FluidOps", 3, 64, "Watdiv100M", "Default", ""), \
("FluidOps", 1, 64, "Watdiv1000M", "Default", ""), \
("FluidOps", 3, 64, "Watdiv1000M", "Default", "")]

runtimes_for_stores(bm_tuples, querytimeout)


### 8. Ontoforce

querytimeout = 1200
bm_tuples = [ ("Blazegraph", 1, 64, "Ontoforce", "Optimized", ""), \
("GraphDB", 1, 64, "Ontoforce", "Optimized", ""), \
("ES", 1, 64, "Ontoforce", "Default", ""), \
("Virtuoso", 1, 64, "Ontoforce", "Optimized", ""), \
("Virtuoso", 1, 64, "Ontoforce", "Optimized", "VWall"), \
("Virtuoso", 1, 32, "Ontoforce", "Optimized", ""), \
("Virtuoso", 1, 32, "Ontoforce", "Optimized", "VWall"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "0"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "1"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "2"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "AWS1"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "AWS2"), \
("Virtuoso", 3, 64, "Ontoforce", "Optimized", "AWS3"), \
("Fuseki", 1, 64, "Ontoforce", "Default", ""), \
("FluidOps", 1, 64, "Ontoforce", "Default", "1"), \
("FluidOps", 1, 64, "Ontoforce", "Default", "2"), \
("FluidOps", 1, 64, "Ontoforce", "Default", "3"), \
("FluidOps", 3, 64, "Ontoforce", "Default", "1"), \
("FluidOps", 3, 64, "Ontoforce", "Default", "2")]
#omitted default 3 since all queries failed

runtimes_for_stores(bm_tuples, querytimeout)


## 3. Query Runtime dataframe from results file: Tryout

* for two simulations (early on in bm cycle) the log files were not stored:
    - GraphDB,1,32,Watdiv1000M,Default
    - ES     ,1,32,Watdiv100M ,Default
    
* Let's manually convert them to the queryruntimes.csv format



bm_tuple1 = ('GraphDB',1,32,'Watdiv1000M','Default','')
results_csv = './GraphDB/results_GraphDB_N1_32_Watdiv1000M_Default.csv'
df_tuple1 = pd.read_csv(results_csv, sep=',', skiprows=34)

interesting_cols = ['Operation', 'Average Runtime (Arithmetic)']
df_tuple1 = df_tuple1[interesting_cols]

df_tuple1['query_name'] = df_tuple1['Operation'].apply(lambda op: op[len('templated/'):])
df_tuple1['median_runtime'] = df_tuple1['Average Runtime (Arithmetic)']
df_tuple1['thread_type'] = 'stress'

df_tuple1['success'] = 1
df_tuple1['errors'] = 0
df_tuple1['timeouts'] = 0

interesting_cols = ['query_name', 'thread_type', 'success', 'errors', 'timeouts', 'median_runtime']
df_tuple2 = df_tuple1[interesting_cols]

new_filename = generateFilenameRuntimesCorrect(bm_tuple1)
df_tuple2.to_csv(new_filename, sep="\t", header=True, index=False)

bm_tuple1 = ('ES',1,32,'Watdiv100M','Default','')
results_csv = './ES/results_ES_N1_32_Watdiv100M_Default.csv'

interesting_cols = ['Operation', 'Average Runtime (Arithmetic)']
df_tuple1 = df_tuple1[interesting_cols]

df_tuple1['query_name'] = df_tuple1['Operation'].apply(lambda op: op[len('templated/'):])
df_tuple1['median_runtime'] = df_tuple1['Average Runtime (Arithmetic)']
df_tuple1['thread_type'] = 'stress'

df_tuple1['success'] = 1
df_tuple1['errors'] = 0
df_tuple1['timeouts'] = 0

interesting_cols = ['query_name', 'thread_type', 'success', 'errors', 'timeouts', 'median_runtime']
df_tuple2 = df_tuple1[interesting_cols]

new_filename = generateFilenameRuntimesCorrect(bm_tuple1)
df_tuple2.to_csv(new_filename, sep="\t", header=True, index=False)
